# back/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        }
    },
    "paths": {
...
        "/tickets": {
            "post": {
                "description": "Crée un billet de tombola pour un élève ou un parent.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tickets"
                ],
                "summary": "Créer un nouveau ticket",
                "parameters": [
                    {
                        "description": "Ticket à créer",
                        "name": "ticket",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Ticket"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Ticket créé avec succès",
                        "schema": {
                            "$ref": "#/definitions/models.Ticket"
                        }
                    },
                    "400": {
                        "description": "Requête invalide",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Erreur interne du serveur",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/tickets/{id}": {
            "get": {
                "description": "Récupère un ticket spécifique par son ID.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tickets"
                ],
                "summary": "Récupérer un ticket par ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Ticket ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ticket trouvé",
                        "schema": {
                            "$ref": "#/definitions/models.Ticket"
                        }
                    },
                    "400": {
                        "description": "ID de ticket invalide",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Ticket non trouvé",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Supprime un ticket spécifique par son ID.",
                "tags": [
                    "Tickets"
                ],
                "summary": "Supprimer un ticket par ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Ticket ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ticket supprimé avec succès",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "ID de ticket invalide",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Ticket non trouvé",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Erreur lors de la suppression du ticket",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/tombola/{tombola_id}/tickets": {
            "get": {
                "description": "Récupère tous les tickets associés à une tombola spécifique.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tickets"
                ],
                "summary": "Récupérer tous les tickets pour une tombola",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Tombola ID",
                        "name": "tombola_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste des tickets",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Ticket"
                            }
                        }
                    },
                    "400": {
                        "description": "ID de tombola invalide",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Erreur lors de la récupération des tickets",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.Ticket": {
            "type": "object",
            "required": [
                "kermesse_id",
                "student_id",
                "ticket_number",
                "tombola_id"
            ],
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "kermesse_id": {
                    "description": "Référence vers la Kermesse",
                    "type": "integer"
                },
                "purchased_at": {
                    "description": "Date d'achat des tickets",
                    "type": "string"
                },
                "student_id": {
                    "description": "Référence vers un Student",
                    "type": "integer"
                },
                "ticket_number": {
                    "description": "Numéro unique du ticket",
                    "type": "string"
                },
                "tombola_id": {
                    "description": "Référence vers la Tombola",
                    "type": "integer"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "Bearer token",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}